Split GeoServerSyncService into per-resource syncers
The 934-line god class becomes a thin coordinator (168 lines) delegating to a new sync/ package: WorkspaceSyncer, StoreSyncer, StyleSyncer, LayerSyncer, sharing bookkeeping via BaseSyncer. Each per-workspace sync method is also decomposed internally — the original 122-line sync_layers_for_workspace (7 mixed concerns) is now three methods (orchestrate / upsert / delete-stale), largest is 76 lines with at most two concerns each.

Fixes two test files that monkey-patched private fetch methods directly on the GeoServerSyncService instance (service._get_geoserver_workspaces = MagicMock(...)) — now a no-op since the coordinator no longer fetches itself. Retargeted 14 occurrences to the syncer instances (service._workspace_syncer._fetch_remote, etc.) across test_geodata_engine_service.py and catalog_api's test_v1_api.py.

Adds 8 new tests proving each syncer works in isolation with a mocked client, per the issue's ask. Public method surface on GeoServerSyncService is unchanged. Full suite: 763 passed (was 755).
